Moving the parallel_traversal algorithm in to its own file.
