+++ nano/node/lmdb/account_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/account_store.hpp>
 #include <nano/node/lmdb/lmdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::lmdb::account_store::account_store (nano::mdb_store & store_a) :
 	store (store_a){};
