+++ nano/node/lmdb/confirmation_height_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/confirmation_height_store.hpp>
 #include <nano/node/lmdb/lmdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::lmdb::confirmation_height_store::confirmation_height_store (nano::mdb_store & store) :
 	store{ store }
 {
