+++ nano/node/lmdb/final_vote_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/final_vote_store.hpp>
 #include <nano/node/lmdb/lmdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::lmdb::final_vote_store::final_vote_store (nano::mdb_store & store) :
 	store{ store } {};
