+++ nano/node/lmdb/pending_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/lmdb.hpp>
 #include <nano/node/lmdb/pending_store.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::lmdb::pending_store::pending_store (nano::mdb_store & store) :
 	store{ store } {};
