+++ nano/node/lmdb/pruned_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/lmdb.hpp>
 #include <nano/node/lmdb/pruned_store.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::lmdb::pruned_store::pruned_store (nano::mdb_store & store_a) :
 	store{ store_a } {};
