+++ nano/node/rocksdb/block_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/rocksdb/block_store.hpp>
 #include <nano/node/rocksdb/rocksdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 namespace nano
 {
 /**
