+++ nano/node/rocksdb/confirmation_height_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/rocksdb/confirmation_height_store.hpp>
 #include <nano/node/rocksdb/rocksdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::rocksdb::confirmation_height_store::confirmation_height_store (nano::rocksdb_store & store) :
 	store{ store }
 {
