+++ nano/node/rocksdb/frontier_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/rocksdb/frontier_store.hpp>
 #include <nano/node/rocksdb/rocksdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::rocksdb::frontier_store::frontier_store (nano::rocksdb_store & store) :
 	store{ store }
 {
