+++ nano/node/rocksdb/unchecked_store.cpp
@@ -1,6 +1,8 @@
 #include <nano/node/lmdb/unchecked_store.hpp>
 #include <nano/node/rocksdb/rocksdb.hpp>
 
+#include <nano/secure/parallel_traversal.hpp>
+
 nano::rocksdb::unchecked_store::unchecked_store (nano::rocksdb_store & store_a) :
 	store (store_a){};
