+++ nano/secure/store_partial.hpp
@@ -11,18 +11,8 @@
 
 #include <thread>
 
-class store_partial;
-namespace
-{
-template <typename T>
-void parallel_traversal (std::function<void (T const &, T const &, bool const)> const & action);
-}
-
 namespace nano
 {
-template <typename Val, typename Derived_Store>
-class store_partial;
-
 /** This base class implements the store interface functions which have DB agnostic functionality. It also maps all the store classes. */
 template <typename Val, typename Derived_Store>
 class store_partial : public store
@@ -145,33 +135,4 @@ protected:
 	virtual int status_code_not_found () const = 0;
 	virtual std::string error_string (int status) const = 0;
 };
-}
-
-namespace
-{
-template <typename T>
-void parallel_traversal (std::function<void (T const &, T const &, bool const)> const & action)
-{
-	// Between 10 and 40 threads, scales well even in low power systems as long as actions are I/O bound
-	unsigned const thread_count = std::max (10u, std::min (40u, 10 * std::thread::hardware_concurrency ()));
-	T const value_max{ std::numeric_limits<T>::max () };
-	T const split = value_max / thread_count;
-	std::vector<std::thread> threads;
-	threads.reserve (thread_count);
-	for (unsigned thread (0); thread < thread_count; ++thread)
-	{
-		T const start = thread * split;
-		T const end = (thread + 1) * split;
-		bool const is_last = thread == thread_count - 1;
-
-		threads.emplace_back ([&action, start, end, is_last] {
-			nano::thread_role::set (nano::thread_role::name::db_parallel_traversal);
-			action (start, end, is_last);
-		});
-	}
-	for (auto & thread : threads)
-	{
-		thread.join ();
-	}
-}
 }
